requirements.txt
Django==2.0
django-filter==1.1.0
django-widget-tweaks==1.4.2
pytz==2018.4
psycopg2==2.7.5
psycopg2-binary